move new reconciler tests to dedicated envtest
