+++ .tools_versions.yaml
@@ -16,3 +16,5 @@ gotestsum: "1.12.0"
 crd-ref-docs: "0.1.0"
 # renovate: datasource=github-releases depName=vektra/mockery
 mockery: "2.45.0"
+# renovate: datasource=github-releases depName=kubernetes-sigs/controller-runtime
+setup-envtest: "0.19.0"
